+++ src/test/java/tests/business/common/BusinessLandingPageTest.java
@@ -1,8 +1,8 @@
-package tests.business;
+package tests.business.common;
 
 import org.testng.Assert;
 import org.testng.annotations.Test;
-import pages.business.BusinessBaseTestClass;
+import pages.business.common.BusinessBaseTestClass;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;