+++ src/test/java/tests/business/manager/BusinessManagerSignUpTest.java
@@ -2,7 +2,7 @@ package tests.business.manager;
 
 import org.testng.Assert;
 import org.testng.annotations.Test;
-import pages.business.BusinessBaseTestClass;
+import pages.business.common.BusinessBaseTestClass;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;